+++ packages/adhere-ui-tablegridlayout/src/types.ts
@@ -33,3 +33,34 @@ export interface ITableGridLayoutProps {
   parity?: boolean;
   data: IDataItem[];
 }
+
+/**
+ * GroupDetail
+ * @description 组的渲染细节
+ */
+export type GroupRenderDetail = Array<{
+  // 一行开始数据的索引
+  startIndex: number;
+  // 一行结束数据的索引
+  endIndex: number;
+}>;
+
+/**
+ * RenderDetail
+ * @description 渲染细节
+ */
+export type RenderDetail = {
+  // 总行数
+  rowCount: number;
+  // 渲染时候的布局
+  layout: 'horizontal' | 'vertical';
+  // 细节
+  detail: Array<{
+    // 组名称
+    name: string;
+    // 总行数
+    rowCount: number;
+    // 细节
+    detail: GroupRenderDetail;
+  }>;
+};